MNT/API: move Handler base classes out of retrieve
 - move HandlerBase, HandlerRegistry, DuplicateHandler to handlers_base
 - re-import to maintain API, but will remove eventually
