+++ filestore/file_writers.py
@@ -1,6 +1,6 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
-from .retrieve import HandlerBase
+from .handlers_base import HandlerBase
 
 import errno
 import six
